Raname ValueTable files
