+++ packages/server/computation_container/client/computation_to_db/value_table.cpp
@@ -1,580 +1,86 @@
+
 #include "value_table.hpp"
 
-#include <openssl/sha.h>
-#include <algorithm>
-#include <chrono>
-#include <cmath>
-#include <unordered_map>
-#include <unordered_set>
 #include "client/computation_to_db/client.hpp"
-#include "job/progress_manager.hpp"
 #include "logging/logger.hpp"
-#include "share/compare.hpp"
-#include "share/share.hpp"
 
 namespace qmpc::ComputationToDb
 {
-using Share = ::Share;
 
-template <class SV = FixedPoint>
-auto toShare(const std::vector<std::string> &vec)
+// nullでない場合にコンテナのbegin()を返す
+std::optional<TableType::iterator> begin(std::optional<TableType> &row)
 {
-    auto sv_vec = std::vector<SV>(vec.begin(), vec.end());
-    auto share_vec = std::vector<Share>(sv_vec.begin(), sv_vec.end());
-    return share_vec;
-}
-template <class SV = FixedPoint>
-auto toShare(const std::vector<std::vector<std::string>> &table)
-{
-    std::vector<std::vector<Share>> &share_table;
-    share_table.reserve(table.size());
-    for (const auto &row : table)
+    if (!row)
     {
-        share_table.emplace_back(toShare<SV>(row));
+        return std::nullopt;
     }
-    return share_table;
+    return row.value().begin();
 }
 
-template <class T>
-std::pair<std::vector<int>, std::vector<int>> intersectionValueIndex(
-    const std::vector<T> &v1, const std::vector<T> &v2
-)
+TableIterator::TableIterator(const std::string &data_id)
+    : data_id(data_id)
+    , piece_id(0)
+    , piece_table(Client::getInstance()->readTable(data_id, piece_id))
+    , itr(begin(piece_table))
 {
-    std::map<T, int, LessFloat> v2_mp;
-    for (unsigned int it = 0; it < v2.size(); ++it)
-    {
-        v2_mp.emplace(v2[it], it);
-    }
-    std::vector<int> v1_it;
-    std::vector<int> v2_it;
-    v1_it.reserve(v1.size());
-    v2_it.reserve(v2.size());
-    for (unsigned int it = 0; it < v1.size(); ++it)
-    {
-        if (v2_mp.find(v1[it]) == v2_mp.end())
-        {
-            continue;
-        }
-        v1_it.emplace_back(it);
-        v2_it.emplace_back(v2_mp[v1[it]]);
-    }
-    return {v1_it, v2_it};
 }
 
-template <class T>
-std::vector<std::pair<int, int>> intersectionSortedValueIndex(
-    const std::vector<T> &sorted_v1, const std::vector<T> &sorted_v2
-)
+RowType TableIterator::operator*()
 {
-    // v1, v2がソートされている必要あり
-    if (sorted_v1.size() == 0 || sorted_v2.size() == 0)
+    if (!itr)
     {
-        return {};
+        // TODO: QMPCのerror logを使う
+        throw std::runtime_error("");
     }
-
-    auto progress_manager = qmpc::Job::ProgressManager::getInstance();
-    const auto job_id = sorted_v1[0].getId().getJobId();
-    auto core_progress =
-        progress_manager->createProgress<qmpc::Job::ProgressIters>(job_id, "hjoin: core", 4);
-
-    int size = sorted_v1.size();
-    int len = sorted_v2.size();
-    // ブロックサイズ(S)と暫定のブロック数(N):N:=size/S
-    // サイズO(N)のBulk比較をO(log(len) + S)回行う
-    int block_size = std::min(100, size);
-    int block_nums = (size + block_size - 1) / block_size;
-
-    // 各ブロックの始点indexとその値
-    std::vector<int> block_it;
-    std::vector<T> block_s;
-    block_it.reserve(block_nums);
-    block_s.reserve(block_nums);
-    for (int i = 0; i < block_nums; ++i)
-    {
-        auto it = i * block_size;
-        block_it.emplace_back(it);
-        block_s.emplace_back(sorted_v1[it]);
-    }
-
-    // parallel binary search
-    // block_itの各要素 i について(v1[i]>=v2[j]) となる最小のjを求める
-    // 求めた j はupperに格納される
-    // 各iとjはv1,v2のブロックの区切りをなす
-    std::vector<int> lower(block_nums, -1);
-    std::vector<int> upper(block_nums, len - 1);
-    int iterated_num = std::log2(len) + 1;
-
-    core_progress->update(1);
-    {
-        auto binary_search_progress = progress_manager->createProgress<qmpc::Job::ProgressIters>(
-            job_id, "hjoin: binary search", iterated_num
-        );
-        for (int progress_i = 0; progress_i < iterated_num; ++progress_i)
-        {
-            std::vector<int> mid_v(block_nums);
-            for (int i = 0; i < block_nums; ++i)
-            {
-                mid_v[i] = (lower[i] + upper[i]) / 2;
-            }
-
-            std::vector<T> target;
-            target.reserve(block_nums);
-            for (const auto &mid : mid_v)
-            {
-                target.emplace_back(sorted_v2[mid]);
-            }
-
-            auto less_eq = qmpc::Share::allLessEq(block_s, target);
-            for (int i = 0; i < block_nums; ++i)
-            {
-                (less_eq[i] ? upper[i] : lower[i]) = mid_v[i];
-            }
-            binary_search_progress->update(progress_i);
-        }
-    }
-
-    core_progress->update(2);
-
-    // parallel linear search
-    // 各ブロックを並列に走査する
-    // sorted_v1は各ブロック i について [begin_its[i].first,  end_its[i].first ) を走査
-    // sorted_v2は各ブロック i について [begin_its[i].second, end_its[i].second) を走査
-    std::vector<std::pair<int, int>> begin_its;
-    std::vector<std::pair<int, int>> end_its;
-    begin_its.reserve(block_nums);
-    end_its.reserve(block_nums);
-
-    for (int i = 0; i < block_nums; ++i)
-    {
-        auto v1_l = block_it[i];
-        auto v1_r = (i + 1 < block_nums) ? block_it[i + 1] : size;
-        auto v2_l = std::max(0, upper[i]);
-        auto v2_r = (i + 1 < block_nums) ? std::max(0, upper[i + 1]) : len;
-        begin_its.emplace_back(v1_l, v2_l);
-        end_its.emplace_back(v1_r, v2_r);
-    }
-    auto now_its = begin_its;
-    core_progress->update(3);
-
-    // intersection_it_list:= v1とv2の積集合のindex
-    std::vector<std::pair<int, int>> intersection_it_list;
-    intersection_it_list.reserve(size);
-    // 並列比較するブロックの数
-    int parallel_nums = now_its.size();
-    std::vector<bool> fin(block_nums, false);
-    {
-        // 進捗最大値は動的に変わりうるため便宜上100に固定
-        auto linear_search_progress = progress_manager->createProgress<qmpc::Job::ProgressIters>(
-            job_id, "hjoin: linear search", 100
-        );
-        while (true)
-        {
-            // l < r を実行したい全てのl, rをcomp_l,comp_rに入れる
-            std::vector<T> comp_l;
-            std::vector<T> comp_r;
-            comp_l.reserve(parallel_nums);
-            comp_r.reserve(parallel_nums);
-            for (int i = 0; i < parallel_nums; ++i)
-            {
-                // 不要な比較を減らすため探索し終えたindexを見ない
-                if (fin[i])
-                {
-                    continue;
-                }
-                auto [i1, i2] = now_its[i];
-                auto [lim1, lim2] = end_its[i];
-                if (i1 == lim1 || i2 == lim2)
-                {
-                    fin[i] = true;
-                    continue;
-                }
-                // lessとgreaterも並列で行う
-                comp_l.emplace_back(sorted_v1[i1]);
-                comp_l.emplace_back(sorted_v2[i2]);
-                comp_r.emplace_back(sorted_v2[i2]);
-                comp_r.emplace_back(sorted_v1[i1]);
-            }
-
-            // comp_lが空 <=> 全てのindexでendまで探索し終えた
-            if (comp_l.empty())
-            {
-                break;
-            }
-
-            // [0th less, 0th greater, 1st less, 1st greater ....]
-            auto comp = qmpc::Share::allLess(comp_l, comp_r);
-            auto comp_it = 0;
-            int min_progress = 100;
-            for (int i = 0; i < parallel_nums; ++i)
-            {
-                if (fin[i])
-                {
-                    continue;
-                }
-                // "==" <=> "not < && not >"
-                if (!comp[comp_it] && !comp[comp_it + 1])
-                {
-                    intersection_it_list.emplace_back(now_its[i]);
-                    ++now_its[i].first;
-                    ++now_its[i].second;
-                }
-                else if (comp[comp_it])
-                {
-                    ++now_its[i].first;
-                }
-                else
-                {
-                    ++now_its[i].second;
-                }
-
-                // 1回の等値比較でlessとgreaterの2つ取り出すので2つ進める
-                comp_it += 2;
-
-                // begin->endに何%移動しているか
-                // v1,v2の大さい方をこのindexでの進捗とする(どちらかが100%で終了のため)
-                auto [b1, b2] = begin_its[i];
-                auto [e1, e2] = end_its[i];
-                auto [i1, i2] = now_its[i];
-                auto progress1 = 100 * (i1 - b1) / (e1 - b1);
-                auto progress2 = 100 * (i2 - b2) / (e2 - b2);
-                min_progress = std::min(min_progress, std::max(progress1, progress2));
-            }
-            linear_search_progress->update(min_progress);
-        }
-    }
-
-    return intersection_it_list;
+    return *(itr.value());
 }
-
-template <class T>
-std::pair<std::vector<int>, std::vector<int>> unionValueIndex(
-    const std::vector<T> &v1, const std::vector<T> &v2
-)
+TableIterator &TableIterator::operator++()
 {
-    std::vector<int> v1_it;
-    v1_it.reserve(v1.size());
-    std::set<T, LessFloat> v1_st;
-    for (unsigned int it = 0; it < v1.size(); ++it)
+    if (!itr || !piece_table)
     {
-        v1_it.emplace_back(it);
-        v1_st.emplace(v1[it]);
+        // TODO: QMPCのerror logを使う
+        throw std::runtime_error("");
     }
-    std::vector<int> v2_it;
-    v2_it.reserve(v2.size());
-    for (unsigned int it = 0; it < v2.size(); ++it)
+
+    ++itr.value();
+    if (itr.value() == piece_table.value().end())
     {
-        if (v1_st.find(v2[it]) != v1_st.end())
-        {
-            continue;
-        }
-        v2_it.emplace_back(it);
+        ++piece_id;
+        piece_table = Client::getInstance()->readTable(data_id, piece_id);
+        itr = begin(piece_table);
     }
-    return {v1_it, v2_it};
+    return *this;
 }
+bool TableIterator::operator!=(const TableIterator &tgt) { return itr != tgt.itr; }
+bool TableIterator::operator==(const TableIterator &tgt) { return !operator!=(tgt); }
 
-std::string writeVJoinTable(
-    const ValueTable &table1,
-    const ValueTable &table2,
-    const std::vector<int> &col1,
-    const std::vector<int> &col2,
-    const std::vector<int> &row1,
-    const std::vector<int> &row2
-)
+ValueTable::ValueTable(const std::string &data_id) : data_id(data_id) {}
+TableIterator ValueTable::begin() const { return TableIterator(data_id); }
+TableIterator ValueTable::end() const { return TableIterator(std::nullopt); }
+
+std::string ValueTable::getDataId() const { return data_id; }
+TableType ValueTable::getTable() const
 {
-    // schemasを構築
-    auto new_schemas = std::vector<std::string>();
-    auto schemas1 = table1.getSchemas();
-    for (const auto &it : col1)
+    TableType table;
+    for (const auto &row : *this)
     {
-        new_schemas.emplace_back(schemas1[it]);
+        table.emplace_back(row);
     }
-
-    // tableを構築
-    auto new_table = std::vector<std::vector<std::string>>();
-    new_table.reserve(row1.size());
-    auto row_dq1 = std::deque<int>(row1.begin(), row1.end());
-    auto row_dq2 = std::deque<int>(row2.begin(), row2.end());
-    auto itr1 = table1.begin();
-    auto itr2 = table2.begin();
-    auto table_idx1 = 0;
-    auto table_idx2 = 0;
-    while (!row_dq1.empty())
-    {
-        // 使用される行になるまでincrement
-        while (table_idx1 < row_dq1.front())
-        {
-            ++table_idx1;
-            ++itr1;
-        }
-        if (itr1 == table1.end())
-        {
-            // TODO: 適切なエラーを返す
-            throw std::runtime_error("");
-        }
-
-        // 行を構築
-        auto new_row = std::vector<std::string>();
-        new_row.reserve(new_schemas.size());
-        auto r1 = *itr1;
-        for (const auto &it : col1)
-        {
-            new_row.emplace_back(r1[it]);
-        }
-        new_table.emplace_back(new_row);
-        ++itr1;
-        ++table_idx1;
-        row_dq1.pop_front();
-    }
-    while (!row_dq2.empty())
-    {
-        // 使用される行になるまでincrement
-        while (table_idx2 < row_dq2.front())
-        {
-            ++table_idx2;
-            ++itr2;
-        }
-        if (itr2 == table2.end())
-        {
-            // TODO: 適切なエラーを返す
-            throw std::runtime_error("");
-        }
-
-        // 行を構築
-        auto new_row = std::vector<std::string>();
-        new_row.reserve(new_schemas.size());
-        auto r2 = *itr2;
-        for (const auto &it : col2)
-        {
-            new_row.emplace_back(r2[it]);
-        }
-
-        new_table.emplace_back(new_row);
-        ++itr2;
-        ++table_idx2;
-        row_dq2.pop_front();
-    }
-
-    // 保存
-    const std::string new_data_id = "tmp";
-    auto db = Client::getInstance();
-    db->writeTable(new_data_id, new_table, new_schemas);
-    return new_data_id;
+    return table;
 }
-std::string writeHJoinTable(
-    const ValueTable &table1,
-    const ValueTable &table2,
-    const std::vector<int> &col1,
-    const std::vector<int> &col2,
-    const std::vector<int> &row1,
-    const std::vector<int> &row2
-)
+std::vector<std::string> ValueTable::getColumn(int column_number) const
 {
-    // schemasを構築
-    auto new_schemas = std::vector<std::string>();
-    auto schemas1 = table1.getSchemas();
-    auto schemas2 = table2.getSchemas();
-    for (const auto &it : col1)
+    std::vector<std::string> ret;
+    for (const auto &row : *this)
     {
-        new_schemas.emplace_back(schemas1[it]);
+        ret.emplace_back(row[column_number]);
     }
-    for (const auto &it : col2)
-    {
-        new_schemas.emplace_back(schemas2[it]);
-    }
-
-    // tableを構築
-    auto new_table = std::vector<std::vector<std::string>>();
-    new_table.reserve(row1.size());
-    auto row_dq1 = std::deque<int>(row1.begin(), row1.end());
-    auto row_dq2 = std::deque<int>(row2.begin(), row2.end());
-    auto itr1 = table1.begin();
-    auto itr2 = table2.begin();
-    auto table_idx1 = 0;
-    auto table_idx2 = 0;
-    while (!row_dq1.empty() && !row_dq2.empty())
-    {
-        // 使用される行になるまでincrement
-        while (table_idx1 < row_dq1.front())
-        {
-            ++table_idx1;
-            ++itr1;
-        }
-        while (table_idx2 < row_dq2.front())
-        {
-            ++table_idx2;
-            ++itr2;
-        }
-        if (itr1 == table1.end() || itr2 == table2.end())
-        {
-            // TODO: 適切なエラーを返す
-            throw std::runtime_error("");
-        }
-
-        // 行を構築
-        auto new_row = std::vector<std::string>();
-        new_row.reserve(new_schemas.size());
-        auto r1 = *itr1;
-        auto r2 = *itr2;
-        for (const auto &it : col1)
-        {
-            new_row.emplace_back(r1[it]);
-        }
-        for (const auto &it : col2)
-        {
-            new_row.emplace_back(r2[it]);
-        }
-
-        new_table.emplace_back(new_row);
-        ++itr1;
-        ++itr2;
-        ++table_idx1;
-        ++table_idx2;
-        row_dq1.pop_front();
-        row_dq2.pop_front();
-    }
-
-    // 保存
-    const std::string new_data_id = "tmp";
-    auto db = Client::getInstance();
-    db->writeTable(new_data_id, new_table, new_schemas);
-    return new_data_id;
+    return ret;
 }
-
-ValueTable vjoin(const ValueTable &table1, const ValueTable &table2, int idx1, int idx2)
+std::vector<std::string> ValueTable::getSchemas() const
 {
-    // joinしたschemasのindexリストを構築
-    auto [schemas_it1, schemas_it2] =
-        intersectionValueIndex(table1.getSchemas(), table2.getSchemas());
-
-    // joinしたidsのindexリストを構築
-    auto ids_share1 = toShare(table1.getColumn(idx1 - 1));
-    open(ids_share1);
-    auto ids1 = recons(ids_share1);
-    auto ids_share2 = toShare(table2.getColumn(idx2 - 1));
-    open(ids_share2);
-    auto ids2 = recons(ids_share2);
-    auto [ids_it1, ids_it2] = unionValueIndex(ids1, ids2);
-
-    // tableをjoinして保存
-    auto new_data_id = writeVJoinTable(table1, table2, schemas_it1, schemas_it2, ids_it1, ids_it2);
-    return ValueTable(new_data_id);
-}
-
-ValueTable hjoin(const ValueTable &table1, const ValueTable &table2, int idx1, int idx2)
-{
-    // joinしたschemasのindexリストを構築
-    auto schemas_it1 = std::vector<int>(table1.getSchemas().size());
-    auto schemas_it2 = std::vector<int>();
-    std::iota(schemas_it1.begin(), schemas_it1.end(), 0);
-    auto size2 = table2.getSchemas().size();
-    for (size_t i = 0; i < size2; ++i)
-    {
-        if (static_cast<int>(i) == idx2 - 1)
-        {
-            continue;
-        }
-        schemas_it2.emplace_back(i);
-    }
-
-    // joinしたidsのindexリストを構築
-    auto ids_share1 = toShare(table1.getColumn(idx1 - 1));
-    open(ids_share1);
-    auto ids1 = recons(ids_share1);
-    auto ids_share2 = toShare(table2.getColumn(idx2 - 1));
-    open(ids_share2);
-    auto ids2 = recons(ids_share2);
-    auto [ids_it1, ids_it2] = intersectionValueIndex(ids1, ids2);
-
-    // tableをjoinして保存
-    auto new_data_id = writeHJoinTable(table1, table2, schemas_it1, schemas_it2, ids_it1, ids_it2);
-    return ValueTable(new_data_id);
-}
-
-ValueTable hjoinShare(const ValueTable &table1, const ValueTable &table2, int idx1, int idx2)
-{
-    // joinしたschemasのindexリストを構築
-    auto schemas_it1 = std::vector<int>(table1.getSchemas().size());
-    auto schemas_it2 = std::vector<int>();
-    std::iota(schemas_it1.begin(), schemas_it1.end(), 0);
-    auto size2 = table2.getSchemas().size();
-    for (size_t i = 0; i < size2; ++i)
-    {
-        if (static_cast<int>(i) == idx2 - 1)
-        {
-            continue;
-        }
-        schemas_it2.emplace_back(i);
-    }
-
-    // joinしたidsのindexリストを構築
-    auto ids_share1 = toShare(table1.getColumn(idx1 - 1));
-    auto ids_share2 = toShare(table2.getColumn(idx2 - 1));
-    // NOTE: 返り値をpairにしたい
-    auto ids_it_list = intersectionSortedValueIndex(ids_share1, ids_share2);
-    std::vector<int> ids_it1;
-    std::vector<int> ids_it2;
-    ids_it1.reserve(ids_it_list.size());
-    ids_it2.reserve(ids_it_list.size());
-    for (const auto &[it1, it2] : ids_it_list)
-    {
-        ids_it1.emplace_back(it1);
-        ids_it2.emplace_back(it2);
-    }
-
-    // tableをjoinして保存
-    auto new_data_id = writeHJoinTable(table1, table2, schemas_it1, schemas_it2, ids_it1, ids_it2);
-    return ValueTable(new_data_id);
-}
-
-auto parseRead(
-    const std::vector<ValueTable> &values,
-    const std::vector<int> &join,
-    const std::vector<int> &index
-)
-{
-    auto joinFunc = [&](auto &&f, const ValueTable &t, unsigned int it = 0)
-    {
-        if (it == join.size())
-        {
-            return t;
-        }
-        if (join[it] == 0)
-        {
-            return f(f, hjoin(t, values[it + 1], index[0], index[it + 1]), it + 1);
-        }
-        if (join[it] == 1)
-        {
-            return f(f, vjoin(t, values[it + 1], index[0], index[it + 1]), it + 1);
-        }
-        return f(f, hjoinShare(t, values[it + 1], index[0], index[it + 1]), it + 1);
-    };
-    return joinFunc(joinFunc, values[0]);
-}
-
-// tableデータを結合して取り出す
-ValueTable readTable(const managetocomputation::JoinOrder &table)
-{
-    // requestからデータ読み取り
-    auto size = table.join().size();
-    std::vector<int> join;
-    join.reserve(size);
-    for (const auto &j : table.join())
-    {
-        join.emplace_back(j);
-    }
-    std::vector<int> index;
-    index.reserve(size);
-    for (const auto &j : table.index())
-    {
-        index.emplace_back(j);
-    }
-    std::vector<ValueTable> tables;
-    tables.reserve(size + 1);
-    for (const auto &data_id : table.dataids())
-    {
-        tables.emplace_back(data_id);
-    }
-    return parseRead(tables, join, index);
+    return Client::getInstance()->readSchema(data_id);
 }
 
 }  // namespace qmpc::ComputationToDb
+
