+++ packages/server/computation_container/test/integration_test/computation_test/BUILD
@@ -14,7 +14,7 @@ cc_library(
         "//test/integration_test:math_test",
         "//test/integration_test:matrix_test",
         "//test/integration_test:read_triple_test",
-        "//test/integration_test:value_table_test",
+        "//test/integration_test:join_table_test",
         "//config_parse:config_parse",
         "//server/computation_to_computation_container:server",
         "//server/manage_to_computation_container:server",
