+++ packages/server/computation_container/test/unit_test/BUILD
@@ -123,7 +123,6 @@ cc_test(
     deps = [
         "@gtest//:main",
         "//client/computation_to_db:client",
-        "//client/computation_to_db:valuetable",
         "@proto//common_types:common_types_cc_proto",
     ],
     linkopts = [
@@ -140,7 +139,7 @@ cc_test(
     copts = ["-Iexternal/gtest/include"],
     deps = [
         "@gtest//:main",
-        "//client/computation_to_db:valuetable_new",
+        "//client/computation_to_db:valuetable",
     ],
     linkopts = [
         "-lstdc++fs"
